requirements.txt
aiosqlite==0.19.0
asgi-ratelimit==0.10.0
fastapi==0.104.0
httpx==0.25.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1
sqlalchemy[asyncio]==2.0.22
uvicorn[standard]==0.23.2
